feat(app): home module Creating the Home module, which will be the main workspace of the application after the login.
This also creates the main cards components, to showcase the user's cards in a carousel, and a big card showing the selected one.

A couple of flows were defined to connect both the Card and CardCarousel components, so that they can communicate through the HomePage.

A main NavBar was also created to be included in every page inside the Home module.