feat(app): not found and loading overlay Creating the basic Not Found page, which will be the point of redirection every time the user encounters an unresolved route.
Creating the Session service, which will be responsible for keeping the active session of the user, in a very simple and mocked fashion. It also contains a canActivate method, which will be responsible for evaluating if the user can enter the Home module.

Also, added the loading overlay to the Login page, which will be shown while the user is logging in.